pyproject.toml
[project]
name = "time-traveler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "attrs==25.3.0",
    "beautifulsoup4==4.13.4",
    "certifi==2025.6.15",
    "cffi==1.17.1",
    "charset-normalizer==3.4.2",
    "exceptiongroup==1.3.0",
    "h11==0.16.0",
    "idna==3.10",
    "numpy==2.0.2",
    "outcome==1.3.0.post0",
    "packaging==25.0",
    "pandas==2.3.0",
    "playwright>=1.53.0",
    "pycparser==2.22",
    "pysocks==1.7.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "pytz==2025.2",
    "requests==2.32.4",
    "selenium==4.33.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "soupsieve==2.7",
    "trio==0.30.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.13.2",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "webdriver-manager==4.0.2",
    "websocket-client==1.8.0",
    "wsproto==1.2.0",
]
